Separate endpoints out into their own files
Only done for the larger view files
