+++ frontstage/views/register/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+
+register_bp = Blueprint('register_bp', __name__,
+                        static_folder='static', template_folder='templates/register')
