+++ frontstage/views/secure_messaging/create_message.py
@@ -1,0 +1,71 @@
+import json
+import logging
+
+from flask import render_template, request
+from frontstage.common.authorisation import jwt_authorization
+from structlog import wrap_logger
+
+from frontstage import app
+from frontstage.common.api_call import api_call
+from frontstage.exceptions.exceptions import ApiError
+from frontstage.models import SecureMessagingForm
+from frontstage.views.secure_messaging.message_get import create_headers, get_message, message_get
+from frontstage.views.secure_messaging import secure_message_bp
+
+
+logger = wrap_logger(logging.getLogger(__name__))
+
+
+@secure_message_bp.route('/create-message', methods=['GET', 'POST'])
+@jwt_authorization(request)
+def create_message(session):
+    party_id = session['party_id']
+    form = SecureMessagingForm(request.form)
+    if request.method == 'POST' and form.validate():
+        is_draft = form.save_draft.data
+        sent_message = send_message(party_id, is_draft)
+
+        # If draft was saved retrieve the saved draft
+        if is_draft:
+            logger.info('Draft sent successfully', message_id=sent_message['msg_id'], party_id=party_id)
+            return message_get('DRAFT', sent_message['msg_id'])
+
+        return render_template('secure-messages/message-success-temp.html', _theme='default')
+
+    else:
+        if form['thread_message_id'].data:
+            message = get_message(form['thread_message_id'].data, 'INBOX', party_id)
+        else:
+            message = {}
+        return render_template('secure-messages/secure-messages-view.html', _theme='default',
+                               form=form, errors=form.errors, message=message.get('message', {}))
+
+
+def send_message(party_id, is_draft):
+    logger.debug('Attempting to send message', party_id=party_id)
+    form = SecureMessagingForm(request.form)
+
+    headers = create_headers()
+    endpoint = app.config['SEND_MESSAGE_URL']
+    subject = form['subject'].data if form['subject'].data else form['hidden_subject'].data
+    message_json = {
+        'msg_from': party_id,
+        'subject': subject,
+        'body': form['body'].data,
+        'thread_id': form['thread_id'].data
+    }
+
+    # If message has previously been saved as a draft add through the message id
+    if form["msg_id"].data:
+        message_json["msg_id"] = form['msg_id'].data
+    response = api_call('POST', endpoint, parameters={"is_draft": is_draft},
+                        json=message_json, headers=headers)
+
+    if response.status_code != 200:
+        logger.debug('Failed to send message', party_id=party_id)
+        raise ApiError(response)
+    sent_message = json.loads(response.text)
+
+    logger.info('Secure message sent successfully',
+                message_id=sent_message['msg_id'], party_id=party_id)
+    return sent_message
