+++ frontstage/views/sign_in/__init__.py
@@ -1,0 +1,4 @@
+from flask import Blueprint
+
+
+sign_in_bp = Blueprint('sign_in_bp', __name__, static_folder='static', template_folder='frontstage/templates/sign-in')
